+++ package.json
@@ -1,7 +1,7 @@
 {
   "engines": {
-    "node": "11.2.0",
-    "yarn": "1.12.3"
+    "node": "10.14.1",
+    "yarn": "1.16.0"
   },
   "scripts": {
     "start": "NODE_ENV=production node build/index.js",
