+++ src/components/icons/Crossmark.tsx
@@ -1,24 +1,13 @@
 import * as React from 'react'
 
 export const Crossmark = () => (
-  <svg
-    width="16"
-    height="16"
-    viewBox="0 0 16 16"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M8 16C12.4183 16 16 12.4183 16 8C16 3.58172 12.4183 0 8 0C3.58172 0 0 3.58172 0 8C0 12.4183 3.58172 16 8 16Z"
-      fill="#FF8A80"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M8.70211 3.49219H7.28611L7.59811 9.56419H8.40211L8.70211 3.49219ZM7.21411 11.2802C7.21411 11.7122 7.56211 12.0602 7.99411 12.0602C8.42611 12.0602 8.78611 11.7122 8.78611 11.2802C8.78611 10.8482 8.42611 10.4882 7.99411 10.4882C7.56211 10.4882 7.21411 10.8482 7.21411 11.2802Z"
-      fill="white"
-    />
+  <svg width="16" height="16" xmlns="http://www.w3.org/2000/svg">
+    <g fill="none" fillRule="evenodd">
+      <circle fill="#FF8A80" cx="8" cy="8" r="8" />
+      <path
+        fill="#FFF"
+        d="M12 4.889L8.889 8 12 11.111l-.889.889L8 8.889 4.889 12 4 11.111 7.112 8 4 4.89 4.889 4 8 7.11 11.11 4z"
+      />
+    </g>
   </svg>
 )
