+++ src/pages/OfferNew/Perils/types.ts
@@ -1,0 +1,8 @@
+export interface Peril {
+  title: string
+  description: string
+  covered: string[]
+  exceptions: string[]
+  info?: string
+  icon: JSX.Element
+}
